feat: Add max button to MarketPost component
The MarketPost component in the browser/src/components directory is updated to add a max button for setting the sell amount to the maximum available amount of the selected resource. This improves the user experience by providing a convenient way to quickly set the sell amount without manually entering the value.